Put split_escaped in config to avoid circular import
