+++ lidarrmetadata/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.3.17dev2'
+__version__ = '0.3.17dev3'
