+++ lidarrmetadata/util.py
@@ -8,7 +8,7 @@ import functools
 from lidarrmetadata import config
 
 # Cache for application
-CACHE = flask_caching.Cache(config=config.CONFIG.CACHE_CONFIG)
+CACHE = flask_caching.Cache(config=config.get_config().CACHE_CONFIG)
 
 
 def cache_or_call(func, *args, **kwargs):
@@ -89,38 +89,6 @@ def map_iterable_values(iterable, func, types=object):
         mapped = original_type(mapped)
 
     return mapped
-
-
-def split_escaped(string, split_char=' ', escape_char='\\'):
-    """
-    Splits escaped string
-
-    :param string: String to split
-    :param split_char: Character to split on. Defaults to single space
-    :param escape_char: Character to escape with. Defaults to \
-    """
-    ret = []
-    current = ''
-    skip = False
-    for i, char in enumerate(string):
-        if skip:
-            skip = False
-            continue
-        elif char == escape_char:
-            current += split_char
-            skip = True
-        elif char == split_char:
-            if current:
-                ret.append(current)
-
-            current = ''
-        else:
-            current += char
-
-    if current:
-        ret.append(current)
-
-    return ret
 
 
 def translate_string(s, table):
